triggerDropdown and actionDropdown files distributed
