--- src/frontend/src/components/ActionDropdown.tsx
@@ -1,376 +1,0 @@
-// Enhanced ActionDropdown component
-import { useState } from "react"
-import { MessageSquare, Mail, FileText } from "lucide-react"
-
-const inputFieldsByApp: Record<
-  string,
-  Array<{
-    key: string
-    label: string
-    placeholder: string
-    type: "text" | "password"
-    required?: boolean
-  }>
-> = {
-  "gmail-send": [
-  {
-    key: "clientId",
-    label: "Client ID",
-    placeholder: "Enter your Gmail OAuth Client ID",
-    type: "text",
-    required: true,
-  },
-  {
-    key: "clientSecret",
-    label: "Client Secret",
-    placeholder: "Enter your Gmail OAuth Client Secret",
-    type: "password",
-    required: true,
-  }
-]
-,
-  "discord-send": [
-  {
-    key: "botToken",
-    label: "Bot Token",
-    placeholder: "Enter your Discord Bot Token",
-    type: "password",
-    required: true,
-  },
-  {
-    key: "applicationId",
-    label: "Application ID",
-    placeholder: "Enter your Discord Application ID",
-    type: "text",
-    required: false,
-  },
-  {
-    key: "guildId",
-    label: "Guild ID",
-    placeholder: "Enter your Discord Server (Guild) ID",
-    type: "text",
-    required: true,
-  },
-  {
-    key: "channelId",
-    label: "Channel ID",
-    placeholder: "Enter the Channel ID to send the message to",
-    type: "text",
-    required: true,
-  }
-]
-,
-  "telegram-send": [
-  {
-    key: "botToken",
-    label: "Bot Token",
-    placeholder: "Enter your Telegram Bot Token",
-    type: "password",
-    required: true,
-  },
-  {
-    key: "chatId",
-    label: "Chat ID",
-    placeholder: "Enter the Chat ID to send the message to",
-    type: "text",
-    required: true,
-  }
-]
-,
-  "notion-create": [
-  {
-    key: "integrationToken",
-    label: "integrationToken",
-    placeholder: "Enter your Notion integrationToken",
-    type: "password",
-    required: true,
-  },
-  {
-    key: "databaseId",
-    label: "Database ID",
-    placeholder: "Enter the Notion Database ID",
-    type: "text",
-    required: true,
-  },
-  {
-    key: "pageID",
-    label: "Page ID",
-    placeholder: "Enter the ID for the new page",
-    type: "text",
-    required: true,
-  }
-]
-,
-  "webhook-post": [
-  {
-    key: "webhookUrl",
-    label: "Webhook URL",
-    placeholder: "Enter the destination webhook URL",
-    type: "text",
-    required: true,
-  },
-  {
-    key: "payload",
-    label: "Payload",
-    placeholder: "Enter the JSON payload to send",
-    type: "text",
-    required: true,
-  }
-]
-,
-  twitter: [
-  {
-    key: "apiKey",
-    label: "API Key",
-    placeholder: "Enter your Twitter API Key",
-    type: "text",
-    required: true,
-  },
-  {
-    key: "apiSecret",
-    label: "API Key Secret",
-    placeholder: "Enter your Twitter API Key Secret",
-    type: "password",
-    required: true,
-  },
-  {
-    key: "accessToken",
-    label: "Access Token",
-    placeholder: "Enter your Twitter Access Token",
-    type: "password",
-    required: true,
-  },
-  {
-    key: "accessTokenSecret",
-    label: "Access Token Secret",
-    placeholder: "Enter your Twitter Access Token Secret",
-    type: "password",
-    required: true,
-  },
-  {
-    key: "bearerToken",
-    label: "Bearer Token",
-    placeholder: "Enter your Twitter Bearer Token (for API v2)",
-    type: "password",
-    required: false, 
-  }
-]
-,
-  "API": [
-  {
-    key: "url",
-    label: "API URL",
-    placeholder: "Enter the API endpoint URL",
-    type: "text",
-    required: true,
-  },
-  {
-    key: "headers",
-    label: "Headers",
-    placeholder: "Enter request headers in JSON format (optional)",
-    type: "text",
-    required: false,
-  },
-  {
-    key: "body",
-    label: "Body",
-    placeholder: "Enter request body for POST/PUT (optional)",
-    type: "text",
-    required: false,
-  }
-]
-,
-  "AI": [
-  {
-    key: "apiKey",
-    label: "AI API Key",
-    placeholder: "Enter your AI service API key (e.g., OpenAI)",
-    type: "password",
-    required: true,
-  },
-  {
-    key: "prompt",
-    label: "Prompt",
-    placeholder: "Enter your input prompt for the AI model",
-    type: "text",
-    required: true,
-  }
-]
-}
-
-const exportEventsByAction: Record<string, Array<{ value: string; label: string; icon: any }>> = {
-  "Send Gmail notification": [
-    { value: "status", label: "Delivery Status", icon: Mail },
-    { value: "messageId", label: "Message ID", icon: Mail },
-  ],
-  "Send custom email": [
-    { value: "subject", label: "Subject Sent", icon: Mail },
-    { value: "timestamp", label: "Timestamp", icon: Mail },
-  ],
-  "Send message": [
-    { value: "messageId", label: "Message ID", icon: MessageSquare },
-    { value: "channel", label: "Channel Sent", icon: MessageSquare },
-  ],
-  "Create page": [
-    { value: "pageId", label: "Page ID", icon: FileText },
-    { value: "createdTime", label: "Creation Time", icon: FileText },
-  ],
-  // Add more for others like Telegram, Notion, Webhook, Twitter, etc.
-}
-
-
-// Add action options for each app
-const actionDropdownOptions: Record<string, string[]> = {
-  "gmail-send": ["Send Gmail notification", "Send custom email", "Send alert email"],
-  "discord-send": ["Send message", "Send alert", "Send notification"],
-  "telegram-send": ["Send message", "Send alert", "Pin message"],
-  "notion-create": ["Create page", "Log to database", "Create task"],
-  "webhook-post": ["Send custom payload", "Trigger external service", "Forward data"],
-  "twitter": ["Send Tweet","Reply to Tweet","Send Tweet with Hashtags"],
-  "AI": ["Generate text","Summarize input","Extract information"],
-  "API": ["Call external API","Send data to endpoint","Trigger webhook"]
-}
-
-export default function ActionDropdown({
-  isOpen,
-  onSave,
-  onCancel,
-  appType,
-}: {
-  isOpen: boolean
-  onSave: (formData: { [key: string]: string }) => void
-  onCancel: () => void
-  appType?: string
-}) {
-  const [credentials, setCredentials] = useState<Record<string, string>>({})
-  const [selectedAction, setSelectedAction] = useState<string>("")
-  const [selectedExport, setSelectedExport] = useState<string>("")
-
-  // Use the full appType (e.g., "gmail-send", "discord-send") instead of splitting
-  const credentialFields = appType ? inputFieldsByApp[appType] || [] : []
-  const actionOptions = appType ? actionDropdownOptions[appType] || [] : []
-  const exportOptions = selectedAction ? exportEventsByAction[selectedAction] || [] : []
-
-  const isFormValid =
-  selectedAction &&
-  selectedExport &&
-  (credentialFields.length === 0 ||
-    credentialFields.every((field) => !field.required || credentials[field.key]?.trim()))
-
-
-  const handleCredentialChange = (key: string, value: string) => {
-    setCredentials((prev) => ({ ...prev, [key]: value }))
-  }
-
-  const handleSave = () => {
-    if (isFormValid) {
-      onSave({
-  ...credentials,
-  event: selectedAction,
-  export: selectedExport,
-})
-    }
-  }
-
-  const handleCancel = () => {
-    setCredentials({})
-    setSelectedAction("")
-    onCancel()
-  }
-
-  if (!isOpen) return null
-
-  return (
-    <div className="mt-3 p-4 bg-[#1b1f2a] border border-[#3a3f52] rounded-lg space-y-4">
-      <div>
-        <h3 className="text-sm font-medium text-white mb-2">Configure Action</h3>
-        <p className="text-xs text-[#9b9bab] mb-3">Select an action and enter required credentials</p>
-      </div>
-
-      <div className="space-y-3">
-        {/* Action Selection */}
-        <div className="flex flex-col gap-2">
-          <label className="text-sm font-medium text-[#c5c5d2]">Select Action</label>
-          <select
-            value={selectedAction}
-            onChange={(e) => {
-                setSelectedAction(e.target.value)
-                setSelectedExport("") // Reset export when action changes
-                }}
-            className="w-full px-3 py-2 bg-[#2a2e3f] border border-[#3a3f52] rounded-md text-white focus:outline-none focus:border-[#6d3be4]"
-          >
-            <option value="" disabled>
-              Choose an action
-            </option>
-            {actionOptions.map((action) => (
-              <option key={action} value={action}>
-                {action}
-              </option>
-            ))}
-          </select>
-        </div>
-
-        {/* Credential Fields */}
-        {credentialFields.map((field) => (
-          <div className="flex flex-col gap-2" key={field.key}>
-            <div className="flex items-center gap-2">
-              <label className="text-sm font-medium text-[#c5c5d2]">{field.label}</label>
-              {field.required && (
-                <span className="bg-red-500/20 text-red-400 text-xs px-1.5 py-0.5 rounded-full">*</span>
-              )}
-            </div>
-            <input
-              type={field.type}
-              value={credentials[field.key] || ""}
-              onChange={(e) => handleCredentialChange(field.key, e.target.value)}
-              placeholder={field.placeholder}
-              className="w-full px-3 py-2 bg-[#2a2e3f] border border-[#3a3f52] rounded-md text-white focus:outline-none focus:border-[#6d3be4]"
-            />
-          </div>
-        ))}
-        {/* Export Options Dropdown */}
-{selectedAction && (
-  <div className="flex flex-col gap-2 pt-3">
-    <label className="text-sm font-medium text-[#c5c5d2]">Select Export Option</label>
-    {exportOptions.length > 0 ? (
-      <select
-        value={selectedExport}
-        onChange={(e) => setSelectedExport(e.target.value)}
-        className="w-full px-3 py-2 bg-[#2a2e3f] border border-[#3a3f52] rounded-md text-white focus:outline-none focus:border-[#6d3be4]"
-      >
-        <option value="" disabled>
-          Choose export value
-        </option>
-        {exportOptions.map((option) => (
-          <option key={option.value} value={option.value}>
-            {option.label}
-          </option>
-        ))}
-      </select>
-    ) : (
-      <div className="text-sm text-[#9b9bab] p-3 bg-[#2a2e3f] rounded-lg">
-        No export options available for this action
-      </div>
-    )}
-  </div>
-)}
-      </div>
-
-      <div className="flex justify-between gap-2 pt-2">
-        <button
-          onClick={handleCancel}
-          className="px-3 py-1 text-sm font-medium text-[#9b9bab] bg-[#2a2e3f] border border-[#3a3f52] rounded-md hover:bg-[#3a3f52]"
-        >
-          Cancel
-        </button>
-        <button
-          onClick={handleSave}
-          disabled={!isFormValid}
-          className="px-3 py-1 text-sm font-medium text-white bg-[#6d3be4] border border-transparent rounded-md hover:bg-[#5a2fc7] disabled:opacity-50 disabled:cursor-not-allowed"
-        >
-          Save Configuration
-        </button>
-      </div>
-    </div>
-  )
-}
